формат даты в ленте новостей

+++ app/src/main/java/com/fktimp/news/models/VKWallModel.kt
@@ -4,7 +4,7 @@ package com.fktimp.news.models
 data class VKWallPost(
     val id: Int = 0,
     val owner_id: Int = 0,
-    val date: Int = 0,
+    val date: Long = 0,
     val text: String = "",
     val reply_post_id: Int = 0
 )
